requirements.txt
## Python version: 3.11.3

torch==2.6.0
numpy==1.25.0
tqdm==4.65.0
transformers==4.33.2
lightning==2.0.9
pillow==9.5.0